requirements.txt
aiogram>=3.13.0
aiohttp>=3.9.0
aiosqlite>=0.19.0
beautifulsoup4>=4.12.0
apscheduler>=3.10.0
aiofiles>=24.1.0
pyyaml>=6.0
